fix: resolve VBox boot hang and 0Eh DMA boundary error
- Implement 16KB chunked disk reads in stage2.asm to avoid crossing 64KB boundaries.
- Add dynamic sector factor detection (512 vs 2048 bytes) for ISO/HDD compatibility.
- Fix stack leak in inode search loop.
- Enforce strict 2048-byte file alignment in format_carleyfs.cpp.
- Improve bootloader diagnostics with string-based error messages and hex codes.

+++ meta/format_carleyfs.cpp
@@ -65,13 +65,14 @@ int main(int argc, char **argv) {
             start = INITRD_SECTOR;
         }
 
-        // ALINEACIÓN CRÍTICA: Forzamos alineación a 2048 bytes (4 sectores lógicos)
-        // Esto asegura que cada archivo empiece al inicio de un sector físico de CD.
+        // ALINEACIÓN CRÍTICA: Forzamos alineación a 2048 bytes (4 sectores lógicos de 512)
+        // Esto asegura que cada archivo empiece al inicio de un sector físico de CD de 2048 bytes.
+        // El Superbloque (128) y los Inodos (132) ya están alineados a 2048 (128 % 4 == 0, 132 % 4 == 0).
         if (start != INITRD_SECTOR) {
             start = (start + 3) & ~3;
         }
 
-        img.seekp(start * 512, std::ios::beg);
+        img.seekp((long long)start * 512, std::ios::beg);
         img.write(buf.data(), buf.size());
 
         std::strncpy(inodes[sb.num_inodes].name, name, 63);
